+++ noxfile.py
@@ -9,24 +9,21 @@ import nox
 nox.options.default_venv_backend = "uv|virtualenv"
 nox.options.reuse_existing_virtualenvs = True
 
-PY39 = "3.9"
 PY310 = "3.10"
 PY311 = "3.11"
 PY312 = "3.12"
 PY313 = "3.13"
 PY314 = "3.14"
-PY_VERSIONS = [PY39, PY310, PY311, PY312, PY313, PY314]
+PY_VERSIONS = [PY310, PY311, PY312, PY313, PY314]
 PY_DEFAULT = PY_VERSIONS[0]
 PY_LATEST = PY_VERSIONS[-1]
 
-DJ42 = "4.2"
-DJ51 = "5.1"
 DJ52 = "5.2"
 DJ60 = "6.0"
 DJ61 = "6.1"
 DJMAIN = "main"
 DJMAIN_MIN_PY = PY312
-DJ_VERSIONS = [DJ42, DJ51, DJ52, DJ60, DJ61, DJMAIN]
+DJ_VERSIONS = [DJ52, DJ60, DJ61, DJMAIN]
 DJ_LTS = [
     version for version in DJ_VERSIONS if version.endswith(".2") and version != DJMAIN
 ]
@@ -46,20 +43,8 @@ def should_skip(python: str, django: str) -> bool:
         # Django main requires Python 3.12+
         return True
 
-    if django in (DJ60, DJ61) and version(python) < version(PY312):
-        # Django 6.0+ requires Python 3.12+
-        return True
-
-    if django in (DJ51, DJ52) and version(python) < version(PY310):
-        # Django 5.1 and 5.2 require Python 3.10+
-        return True
-
-    if django == DJ42 and version(python) > version(PY312):
-        # Django 4.2 supports Python 3.12 and below
-        return True
-
-    # Django 5.1 supports Python 3.13 and below
-    return django == DJ51 and version(python) > version(PY313)
+    # Django 6.0+ requires Python 3.12+
+    return django in (DJ60, DJ61) and version(python) < version(PY312)
 
 
 @nox.session